pyproject.toml
[project]
name = "interpolation_lib"
version = "0.1.0"
description = "Библиотека для интерполяции изображений и данных"
authors = [
    {name = "IsaenkovDanil", email = "isaenkov.dn@phystech.edu"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["interpolation", "image processing", "numpy", "scipy", "data analysis"]

dependencies = [
    "numpy>=2.0.0,<3.0.0",
    "matplotlib>=3.8.0,<4.0.0",
    "pillow>=10.0.0,<12.0.0",
    "click>=8.1.0,<9.0.0"
]

packages = [{ include = "interpolation_lib", from = "src" }]

[project.scripts]
interpolate = "interpolation_lib.cli:interpolate_image_cli"

[project.urls]
Repository = "https://github.com/IsaenkovDanil/Interpolation_IITP"



[tool.poetry.group.dev.dependencies]
ruff = ">=0.4.0"
pytest = "^8.0"
pytest-cov = "^5.0"
pytest-mock = "^3.14.0"
nox = "^2025.2.9"
nox-poetry = "^1.2.0"

[tool.ruff]
target-version = "py313"
line-length = 88
[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = []
fixable = ["ALL"]
unfixable = []
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"